+++ Editor/Engine/Platform/Windows/Renderer/DirectX12/DX12Viewport.h
@@ -6,8 +6,8 @@ struct DX12Viewport
 	D3D12_RECT scissorRect;
 	Vector Pos;
 	Vector Dim;
-	void Init(u32 x, u32 y, u32 Width, u32 Height);
+	void Init(f32 x, f32 y, f32 Width, f32 Height);
 	
-	void Resize(u32 Width, u32 Height);
+	void Resize(f32 Width, f32 Height);
 	void SetViewport(DX12CommandQueue* Queue);
 };
