+++ Editor/Engine/System/Camera/FreeCamera.h
@@ -9,5 +9,5 @@ void FreeCamera_MButtonDown(void* Parent, Mouse mouse);
 void FreeCamera_MButtonUp(void* Parent, Mouse mouse);
 void FreeCamera_Move(void* Parent, Mouse mouse);
 
-void FreeCamera_KeyDown(void* Parent, u32 KeyCode);
-void FreeCamera_KeyUp(void* Parent, u32 KeyCode);
+void FreeCamera_KeyDown(void* Parent, u64 KeyCode);
+void FreeCamera_KeyUp(void* Parent, u64 KeyCode);
